requirements.txt
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
numpy==1.24.3
arabic-reshaper==3.0.0
python-bidi==0.4.2
streamlit==1.31.0
fonttools==4.38.0 